Aligned panning direction of motion with camera
The direction of motion for panning is now aligned with the heading of the camera. Pushing the joystick forward or backward will now move forward and backward along the camera heading.  Pushing it left or right will now move left or right with respect to camera heading. Previous directions of motion were always North/South and East/West.